remove protected datasets button refactor
